requirements.txt
Django==3.1.7
djangorestframework==3.12.4
mysqlclient==2.0.3
django-mysql==3.11.1
django-cors-headers==3.7.0
PyJWT==2.2.0
Faker==9.3.1